This change fixes an issue where Connections list and the Model TimerModel properties of ServerComponent were being accessed by multiple threads without synchronization.
Both the connection list and the timer model were accessed from connection_ScriptReceived and connection_MessageReceived callbacks which come from a thread created by the Connection class. The connection list is also accessed by the UI thread, and the timer model does not appear to support being accessed by non-UI threads.

With this change, the logic that used to run directly in the connection callbacks is moved to the UI thread so that there aren't concurrency issues.